Restore stored state using the hook
